codegen: compare the fast allocator's operand sentinel as u8
`operand_registers` is a `u8` array whose unassigned sentinel is
UINT8_MAX, so the early-clobber rebinding loop's `== UINT32_MAX` guard
could never fire. The loop only behaved because its inner filter skips
UINT8_MAX slots on its own; make the guard match the sentinel it meant.

+++ src/buster/lib/compiler/codegen/register_allocator_fast.c
@@ -1689,7 +1689,7 @@ MachineStackPlacement machine_fast_placement_build_prepassed(Arena* arena, Machi
                     continue;
                 }
                 u32 destination = operand_registers[slot];
-                if (destination == UINT32_MAX)
+                if (destination == UINT8_MAX)
                 {
                     continue;
                 }
